# docs/swagger.yaml
basePath: /v1/user-services
definitions:
  apperror.Code:
    enum:
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - METHOD_NOT_ALLOWED
    - NOT_ACCEPTABLE
    - REQUEST_TIMEOUT
    - CONFLICT
    - GONE
    - PRECONDITION_FAILED
    - PAYLOAD_TOO_LARGE
    - UNSUPPORTED_MEDIA_TYPE
    - VALIDATION_ERROR
    - UNPROCESSABLE_ENTITY
    - LOCKED
    - FAILED_DEPENDENCY
    - PRECONDITION_REQUIRED
    - TOO_MANY_REQUESTS
    - INTERNAL_ERROR
    - NOT_IMPLEMENTED
    - BAD_GATEWAY
    - SERVICE_UNAVAILABLE
    - TIMEOUT
    - CACHE_MISS
    - CACHE_ERROR
    - MESSAGE_BUS_ERROR
    type: string
    x-enum-varnames:
    - CodeBadRequest
    - CodeUnauthorized
    - CodeForbidden
    - CodeNotFound
    - CodeMethodNotAllowed
    - CodeNotAcceptable
    - CodeRequestTimeout
    - CodeConflict
    - CodeGone
    - CodePrecondFailed
    - CodePayloadTooLarge
    - CodeUnsupportedMedia
    - CodeValidation
    - CodeUnprocessable
    - CodeLocked
    - CodeFailedDependency
    - CodePrecondRequired
    - CodeTooManyRequest
    - CodeInternal
    - CodeNotImplemented
    - CodeBadGateway
    - CodeServiceUnavail
    - CodeTimeout
    - CodeCacheMiss
    - CodeCacheError
    - CodeMessageBus
...
  github_com_modami_user-service_internal_dto.KYCDocumentInput:
    properties:
      doc_type:
        enum:
        - id_card_front
        - id_card_back
        - business_license
        - selfie_with_id
        type: string
      doc_url:
        type: string
    required:
    - doc_type
    - doc_url
    type: object
  github_com_modami_user-service_internal_dto.KYCStatusResponse:
    properties:
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_modami_user-service_internal_dto.RegisterSellerRequest:
    properties:
      bank_account:
        type: string
      bank_name:
        maxLength: 128
        type: string
      business_type:
        enum:
        - individual
        - business
        type: string
      shop_description:
        maxLength: 1000
        type: string
      shop_name:
        maxLength: 255
        minLength: 3
        type: string
      shop_slug:
        maxLength: 255
        minLength: 3
        type: string
      tax_id:
        type: string
    required:
    - business_type
    - shop_name
    - shop_slug
    type: object
...
  github_com_modami_user-service_internal_dto.SellerProfileResponse:
    properties:
      avg_rating:
        type: number
      business_type:
        type: string
      created_at:
        type: string
      kyc_status:
        type: string
      shop_banner_url:
        type: string
      shop_description:
        type: string
      shop_logo_url:
        type: string
      shop_name:
        type: string
      shop_slug:
        type: string
      total_reviews:
        type: integer
      user_id:
        type: string
    type: object
  github_com_modami_user-service_internal_dto.SubmitKYCRequest:
    properties:
      documents:
        items:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.KYCDocumentInput'
        minItems: 1
        type: array
    required:
    - documents
    type: object
...
  github_com_modami_user-service_internal_dto.UpdateSellerProfileRequest:
    properties:
      bank_account:
        type: string
      bank_name:
        maxLength: 128
        type: string
      shop_banner_url:
        type: string
      shop_description:
        maxLength: 1000
        type: string
      shop_logo_url:
        type: string
      shop_name:
        maxLength: 255
        minLength: 3
        type: string
      tax_id:
        type: string
    type: object
...
  response.ErrorBody:
    properties:
      code:
        $ref: '#/definitions/apperror.Code'
      detail:
        type: string
      errors:
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        type: string
    type: object
  response.FieldError:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.Meta:
    properties:
      pagination:
        $ref: '#/definitions/response.Pagination'
      request_id:
        type: string
      timestamp:
        type: integer
    type: object
  response.Pagination:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page:
        type: integer
      per_page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorBody'
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
host: localhost:8086
info:
  contact: {}
  description: User service for the Modami marketplace platform.
  title: Modami User Service API
  version: "1.0"
paths:
...
  /users/{id}/shop:
    get:
      description: Returns a seller's public shop profile
      parameters:
      - description: User ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.SellerProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get shop profile
      tags:
      - Sellers
...
  /users/me/seller/kyc:
    post:
      consumes:
      - application/json
      description: Submit KYC documents for verification
      parameters:
      - description: KYC documents
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.SubmitKYCRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Submit KYC documents
      tags:
      - KYC
  /users/me/seller/kyc/status:
    get:
      description: Returns the KYC verification status for the authenticated seller
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.KYCStatusResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get KYC status
      tags:
      - KYC
  /users/me/seller/profile:
    put:
      consumes:
      - application/json
      description: Update the authenticated seller's profile
      parameters:
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.UpdateSellerProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.SellerProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update seller profile
      tags:
      - Sellers
  /users/me/seller/register:
    post:
      consumes:
      - application/json
      description: Register the authenticated user as a seller
      parameters:
      - description: Seller registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_modami_user-service_internal_dto.RegisterSellerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_modami_user-service_internal_dto.SellerProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Already a seller
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Register as seller
      tags:
      - Sellers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"